Move to function-generated tables to table_factory.py
