+++ py/server/deephaven/__init__.py
@@ -23,7 +23,7 @@ from .csv import write as write_csv
 from .stream.kafka import consumer as kafka_consumer
 from .stream.kafka import producer as kafka_producer
 from .table_factory import empty_table, time_table, merge, merge_sorted, new_table, DynamicTableWriter, input_table, \
-    ring_table
+    ring_table, function_generated_table
 from .replay import TableReplayer
 from ._gc import garbage_collect
 from .dbc import read_sql
