--- py/server/deephaven/util/__init__.py
@@ -1,1 +1,0 @@
-from .function_generated_table import create as create_generated_table
